+++ cmd/agent/main.go
@@ -32,7 +32,7 @@ const (
 	provisionTimeout = 50 * time.Minute
 )
 
-var version = "0.7.5"
+var version = "0.7.6"
 
 func main() {
 	if len(os.Args) > 1 && os.Args[1] == "--version" {